samples/01-QUICKSTART/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
asyncer==0.0.7
azure-ai-inference==1.0.0b8
azure-ai-projects==1.0.0b5
azure-common==1.1.28
azure-core==1.32.0
azure-core-tracing-opentelemetry==1.0.0b11
azure-identity==1.19.0
azure-monitor-opentelemetry==1.6.4
azure-monitor-opentelemetry-exporter==1.0.0b33
azure-search-documents==11.5.2
bidict==0.23.1
certifi==2025.1.31
cffi==1.17.1
chainlit==2.2.0
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
cryptography==44.0.0
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.18
distro==1.9.0
executing==2.2.0
fastapi==0.115.8
filetype==1.2.0
fixedint==0.1.6
googleapis-common-protos==1.66.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.32.0
isodate==0.7.2
jedi==0.19.2
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
Lazify==0.4.0
literalai==0.1.103
marshmallow==3.26.1
matplotlib-inline==0.1.7
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
oauthlib==3.2.2
openai==1.61.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-dbapi==0.51b0
opentelemetry-instrumentation-django==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-flask==0.51b0
opentelemetry-instrumentation-psycopg2==0.51b0
opentelemetry-instrumentation-requests==0.51b0
opentelemetry-instrumentation-urllib==0.51b0
opentelemetry-instrumentation-urllib3==0.51b0
opentelemetry-instrumentation-wsgi==0.51b0
opentelemetry-proto==1.29.0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==5.9.8
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
pywin32==308
pyzmq==26.2.1
requests==2.32.3
requests-oauthlib==2.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
syncer==2.0.3
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.29.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==0.20.0
wcwidth==0.2.13
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
